+++ R/mono_utils.R
@@ -32,9 +32,13 @@
 #'         on the tree. See [plot.MonoClust()] for details.}
 #'       \item{split.order}{Order of the splits with root is 0.}
 #'       \item{inertia_explained}{Percent inertia explained as described in
-#'         Chavent (2007). It is \eqn{1 - (sum(current inertia)/inertial[1])}.}
-#'       \item{alt}{Indicator of an alternative cut yielding the same reduction
-#'         in inertia at that split.}
+#'         Chavent (2007). It is `1 - (sum(current inertia)/inertial[1])`.}
+#'       \item{alt}{A nested tibble of alternate splits at a node. It contains
+#'         `bipartsplitrow` and `bipartsplitcol` with the same meaning above.
+#'         Note that this is only for information purpose. Currently `monoClust`
+#'         does not support choosing an alternate splitting route. Running
+#'         [MonoClust()] with `nclusters = 2` step-by-step can be run if
+#'         needed.}
 #'     }}
 #'   \item{membership}{Vector of the same length as the number of rows in the
 #'     data, containing the value of `frame$number` corresponding to the leaf
@@ -54,7 +58,7 @@
 #' @references
 #' * Chavent, M., Lechevallier, Y., & Briant, O. (2007). DIVCLUS-T: A monothetic
 #' divisive hierarchical clustering method. Computational Statistics & Data
-#' Analysis, 52(2), 687–701. <doi:10.1016/j.csda.2007.03.013>.
+#' Analysis, 52(2), 687-701. <doi:10.1016/j.csda.2007.03.013>.
 #' @seealso [MonoClust()].
 NULL
 
@@ -217,9 +221,11 @@ new_node <- function(number,
                      medoid,
                      loc,
                      split.order = -99L,
-                     alt = FALSE) {
+                     alt = list(
+                       tibble::tibble(bipartsplitrow = numeric(),
+                                      bipartsplitcol = numeric()))) {
 
-  one_row_table <- dplyr::tibble(
+  one_row_table <- tibble::tibble(
     number, var, cut, n,
     inertia, bipartsplitrow,
     bipartsplitcol, inertiadel,